% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{RunScGFT}
\alias{RunScGFT}
\title{Generates synthetic scRNA-seq data in-silico}
\usage{
RunScGFT(object, nsynth, ncpmnts = 1, groups, cells = NULL)
}
\arguments{
\item{object}{A seurat object.}

\item{nsynth}{Number of synthetic cells to be generated.}

\item{ncpmnts}{The number of components to modify during the
synthesis process.}

\item{groups}{A character in the original \code{object} metadata used
for synthesis to capture distinct variation modes
in each cell population, such as cell types or clusters.
All groups are uniformly scaled for desired expansion.}

\item{cells}{An optional argument specifying the subset of cells
to be used for synthesis.}
}
\value{
Returns a combined \code{object} of original and synthetic cells. The
  synthesized cells are integrated into the \code{data} layer, while their
  adjusted counts are integrated into the \code{counts} layer.
  The synthesized cells are labeled by the cell barcodes they originated
  from, suffixed with "_synth". A new \code{synthesized} column is added to
  metadata indicating the synthesis status of each cell, labeled as either
  "yes" or "no".
}
\description{
Leverages the principles of Fourier Transform, to generate
  synthetic cells, while preserving the biological plausibility of the
  original population.
}
\details{
The input \code{object} should be log normalized and contain top
  variable features.
}
