% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{statsScGFT}
\alias{statsScGFT}
\title{synthesized cells QC}
\usage{
statsScGFT(object, groups)
}
\arguments{
\item{object}{A Seurat object that includes synthesized cells, i.e., an object post \link{RunScGFT}.}

\item{groups}{Same character in the original \code{object} metadata used for synthesis by \link{RunScGFT}.}
}
\value{
synthesis accuracy in percentage.
}
\description{
Evaluates the gene expression profile of synthesized cells.
}
\details{
assesses the quality of the synthesis process by calculating the
  proportion of synthesized cells that correctly grouped with their
  corresponding original cells according to the given \code{groups} used for synthesis.
}
