% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgft.R
\docType{package}
\name{scgft}
\alias{scgft}
\title{The scgft package}
\description{
\code{scgft} offers a streamlined computational framework, leveraging
the principles of Fourier Transform, to generate synthetic scRNA-seq data in-silico.
}
\section{Synthesis and evaluation of synthetic cells}{

\itemize{
  \item  \link{RunScGFT}:     Generates synthetic scRNA-seq data in-silico
  \item  \link{statsScGFT}:   Calculates the likelihood that synthesized
  cells will have the same identity as their original counterparts.
}
}

\references{
\enumerate{
  \item  Nima Nouri: scGFT - single-cell RNA-seq data augmentation
  using generative Fourier transformers
}
}
